Fixed photos and album to only work when correct user is calling, finally learned how to spell success....

+++ Api-Photos-Emilia/src/controllers/photos.controller.ts
@@ -6,11 +6,23 @@ import { CreatePhotoData, UpdatePhotoData } from "../types/Photo.types.ts";
 
 // Get all photos
 
-export const index = async (_req: Request, res: Response) => {
+export const index = async (req: Request, res: Response) => {
 	try {
-		const photos = await getPhotos();
+
+		if (!req.token) {
+			throw new Error("Authenticated user does not exist");
+		}
+
+		const userId = Number(req.token.sub);
+
+		if (!userId) {
+			res.status(403).send({ status: "fail", data: { message: "Access forbidden" } });
+			return;
+		}
+
+		const photos = await getPhotos(userId);
+
 		res.status(200).send({ status: "success", data: photos })
-		//userId ska inte va med, fixa!
 	} catch (err) {
 		handlePrismaError(res, err);
 	}
@@ -27,10 +39,21 @@ export const show = async (req: Request, res: Response) => {
 		return;
 	}
 
+	if (!req.token) {
+		throw new Error("Authenticated user does not exist");
+	}
+
+	const userId = Number(req.token.sub);
+
+	if (!userId) {
+		res.status(403).send({ status: "fail", data: { message: "Access forbidden" } });
+		return;
+	}
+
 	try {
-		const photo = await getPhoto(photoId);
+		const photo = await getPhoto(photoId, userId);
 		res.status(200).send({
-			status: "succes", data: {
+			status: "success", data: {
 				id: photo.id,
 				title: photo.title,
 				url: photo.url,
@@ -52,6 +75,11 @@ export const store = async (req: Request, res: Response) => {
 
 	const validatedData = matchedData<CreatePhotoData>(req);
 	const userId = Number(req.token.sub);
+
+	if (!userId) {
+		res.status(403).send({ status: "fail", data: { message: "Access forbidden" } });
+		return;
+	}
 
 	try {
 		const photo = await createPhoto(validatedData, userId);
@@ -80,17 +108,28 @@ export const update = async (req: Request, res: Response) => {
 		return;
 	}
 
+	if (!req.token) {
+		throw new Error("Authenticated user does not exist");
+	}
+
+	const userId = Number(req.token.sub);
+
+	if (!userId) {
+		res.status(403).send({ status: "fail", data: { message: "Access forbidden" } });
+		return;
+	}
+
 	const validatedData = matchedData<UpdatePhotoData>(req);
 
 	try {
-		const photo = await updatePhoto(photoId, validatedData);
+		const photo = await updatePhoto(photoId, userId, validatedData);
 		res.status(200).send({
 			status: "success", data: {
 				title: photo.title,
 				url: photo.url,
 				comment: photo.comment,
-				//user_id: //user id?
-				//id: photo.id,
+				user_id: photo.userId,
+				id: photo.id
 			}
 		});
 	} catch (err) {
@@ -109,7 +148,18 @@ export const destroy = async (req: Request, res: Response) => {
 		return;
 	}
 
-	try {  //Kolla så den raderar länkarna till albumen, men inte själva albumen
+	if (!req.token) {
+		throw new Error("Authenticated user does not exist");
+	}
+
+	const userId = Number(req.token.sub);
+
+	if (!userId) {
+		res.status(403).send({ status: "fail", data: { message: "Access forbidden" } });
+		return;
+	}
+
+	try {
 		await deletePhoto(photoId);
 		res.status(200).send({ status: "success", data: null });
 	} catch (err) {
